+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+- Add `get_list_continuous_queries`, `drop_continuous_query`, and `create_continuous_query` management methods for
+  continuous queries (#681 thx @lukaszdudek-silvair)
 - query() now accepts a bind_params argument for parameter binding (#678 thx @clslgrnc)
 
 ### Changed